Add New-O365OrgContact function and handle creation of O365 organization contact.
The code adds a new PowerShell function called New-O365OrgContact that handles the creation of an Office 365 organization contact. It includes a param block with a single parameter called Source. The function uses the Write-Color function to output a message indicating that a contact is being added. It then tries to create a new mail contact using the New-MailContact cmdlet, passing in the values from the Source parameter. If the creation fails, an error message is displayed. Finally, if the contact is successfully created, it calls the Set-O365OrgContact function to further process the contact.