feat: Add synchronization of O365 contacts
This commit adds the functionality to synchronize O365 contacts with a specified list of source objects. The code now includes the conversion of source objects to contact objects and checks if the contact already exists in O365. If the contact does not exist, a new contact is created. If the contact already exists, the code checks if the contact data is the same and updates it if necessary.